chore: move generic merge code outside merge action
This separate the queuing/merging code and the status reporting from the
conditions to queue/merge the code
