+++ mergify_engine/debug.py
@@ -25,7 +25,7 @@ from mergify_engine import queue
 from mergify_engine import rules
 from mergify_engine import subscription
 from mergify_engine import utils
-from mergify_engine.actions import merge
+from mergify_engine.actions import merge_base
 from mergify_engine.clients import github
 from mergify_engine.clients import http
 from mergify_engine.engine import actions_runner
@@ -216,7 +216,7 @@ def report(
                     pulls, key=lambda v: q.get_config(v)["priority"]
                 ):
                     try:
-                        fancy_priority = merge.PriorityAliases(priority).name
+                        fancy_priority = merge_base.PriorityAliases(priority).name
                     except ValueError:
                         fancy_priority = priority
                     formatted_pulls = ", ".join((f"#{p}" for p in grouped_pulls))
